Don't assume qreal is a double

+++ libgui/knob.cpp
@@ -138,7 +138,7 @@ void Knob::paintEvent( QPaintEvent* ) {
 	if ( 0<dbmax && 0>dbmin )
 		_texts << 0.0;
 	QFont small = font();
-	small.setPointSizeF( qMax( 5.0, font().pointSizeF()/2 ) );
+	small.setPointSizeF( qMax( qreal(5.0), font().pointSizeF()/2 ) );
 	p.setFont( small );
 	foreach( double a, _texts ) {
 		p.save();
